initial-app/ascendas_apps/itsa_backend/requirements.txt
asgiref==3.2.10
boto3==1.14.12
botocore==1.17.63
certifi==2020.6.20
chardet==3.0.4
Django==3.1.1
django-cors-headers==3.5.0
django-ebhealthcheck==1.0.1
django-elastipymemcache==2.0.4
django-environ==0.4.5
django-health-check==3.14.3
django-pymemcache==1.0.0
djangorestframework==3.11.1
docutils==0.15.2
gunicorn==20.0.4
idna==2.10
jmespath==0.10.0
loadtest==0.0.2
mysqlclient==2.0.1
pymemcache==3.4.0
python-dateutil==2.8.1
python-dotenv==0.15.0
pytz==2020.1
PyYAML==5.3.1
requests==2.24.0
s3transfer==0.3.3
six==1.15.0
sqlparse==0.3.1
urllib3==1.25.11
